+++ tf/docs/about/releases.md
@@ -10,6 +10,15 @@ See the app-specific tutorials via `tf.about.corpora`.
 
 ## 10
 
+## 10.2.2
+
+2022-09-08
+
+Changes in the output of Text-Fabric to the console.
+It is detected whether it runs in interactive mode (e.g. Jupyter notebook) or not.
+If not, the display methods of the Jupyter notebook are suppressed, 
+and many outputs are done in plain text instead of HTML.
+
 ## 10.2.1
 
 2022-08-23